requirements.txt
discord.py
python-dotenv
twitchio
aiosqlite